+++ src/layers/core/grid-cell-layer/grid-cell-layer.js
@@ -32,6 +32,7 @@ const DEFAULT_COLOR = [255, 0, 255, 255];
 
 const defaultProps = {
   cellSize: 1000,
+  coverage: 1,
   elevationScale: 1,
   extruded: true,
   fp64: false,
@@ -133,7 +134,7 @@ export default class GridCellLayer extends Layer {
   }
 
   updateUniforms() {
-    const {opacity, extruded, elevationScale, cellSize, lightSettings} = this.props;
+    const {opacity, extruded, elevationScale, cellSize, coverage, lightSettings} = this.props;
     const {viewport} = this.context;
     const {pixelsPerMeter} = viewport.getDistanceScales();
 
@@ -141,7 +142,8 @@ export default class GridCellLayer extends Layer {
       extruded,
       elevationScale,
       opacity,
-      cellSize: cellSize * pixelsPerMeter[0]
+      cellSize: cellSize * pixelsPerMeter[0],
+      coverage
     },
     lightSettings));
   }